refactor(p2p-node): extract message queue
Extract all queue logic into a simple message broker. Also extract the
message retry part of the message queue to a retry helper class, which
is used as a parent for message broker.

In addition, update some types to organize the format of the messages
and remove some redundant data.