feat(p2p-node): add randomness to message retry timeout
There is a fixed interval between p2p nodes and relays, in which the
connection is reset by the relay. By having a fixed exponential method
for retry timeout, we may retry the messages exactly when the
connections are reset, which neutralizes the whole retry mechanism.

Add a bit of randomness to the algorithm, so that these kinds of issues
have much less chance to happen.