test(simple-star): refactor test and add peer discovery postcondition
The current test can only read a limited number of logs, which may not
be suitable for all scenarios. This log limit was added because the
follow flag of docker log data source caused it never to finish.

To bypass this limitation, enable log following; then make the docker
files utilize ENTRYPOINT instead of CMD to act as executables, and exit
test process after a specific timeout, so that the container is shutdown
and all logs can be read.

In addition, remove redundant node creation postcondition and add a new
postcondition for checking if the other node is discovered.